Fixed undefined behavior while generating destination path
Context: sprintf was used to generate destination
directories. A string variable was used both for input
and output. The specification says the result is
undefined. On some systems, the destination directory
ended up in the current path:
./.X (where X is the process ID) was chosen instead of
{path provided by the -d argument}/#test-dir.X

This patch simplifies the logic and fixes the creation of
the destination path.